オーバーロード

+++ src/overloads.ts
@@ -12,6 +12,6 @@ function double(value: any): any {
   }
 }
 
-console.log(double(100));
+console.log(double(200));
 console.log(double('Go '));
 //console.log(double(true));
